Migrated all the executables to the new parsing system.
