+++ src/CMakeLists.txt
@@ -1,5 +1,5 @@
 # define all target executables
-set(TARGETS "vasma" "vme" "dvasm" "vasm2x86_64" "parser")
+set(TARGETS "vasma" "vme" "dvasm" "vasm2x86_64")
 
 # get all source (cpp) files and remove executable specific source files
 file(GLOB src "*.cpp")
