+++ src/vasm2x86_64.cpp
@@ -3,6 +3,7 @@
 
 #include "vm.h"
 #include "program_args.h"
+#include "parser.h"
 
 void program_usage(const char* program_name) {
     std::cerr << "Usage: " << program_name << " [args]" << std::endl;
@@ -26,7 +27,9 @@ int main(int argc, char const *argv[]) {
     }
 
     Program p;
-    p.read_from_file(input_file_path.data());
+    Lexer lexer(input_file_path.data());
+    Parser parser(lexer.tokenize());
+    parser.parse(&p);
 
     // get places for labels
     size_t label_suffix = 0;
